Added automatic language detection for custom words. Currently only English and German supported. Added possibility to write a lowercase and an uppercase word as two different correct guesses (if both words exist).
Added function updateHighscore so addHighscore should only add Highscores and not update old ones anymore.
Added picker view delegate and data source to WordSelectionTableViewController in case we need it later to choose the language.

Known problems: If the word is possible in multiple languages it just takes the first in the list of languages (from Language.allLanguages). Should be fixed by asking the user for the language if multiple Languages are possible.